Switch Newton over to using imports rather than parts
Fixes #1355
