+++ packages/newton/lib/src/clamped_simulation.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of newton;
+import 'simulation.dart';
 
 class ClampedSimulation extends Simulation {
   ClampedSimulation(this.simulation, {
