+++ packages/newton/lib/src/friction_simulation.dart
@@ -1,8 +1,11 @@
-// Copyright (c) 2015 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of newton;
+import 'dart:math' as math;
+
+import 'simulation.dart';
+import 'tolerance.dart';
 
 class FrictionSimulation extends Simulation {
   final double _drag;
