+++ packages/newton/lib/src/gravity_simulation.dart
@@ -1,8 +1,8 @@
-// Copyright (c) 2015 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of newton;
+import 'simulation.dart';
 
 class GravitySimulation extends Simulation {
   final double _x;
