+++ packages/newton/lib/src/simulation.dart
@@ -1,8 +1,8 @@
-// Copyright (c) 2015 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of newton;
+import 'tolerance.dart';
 
 abstract class Simulatable {
   /// The current position of the object in the simulation
