+++ packages/newton/lib/src/tolerance.dart
@@ -1,8 +1,6 @@
-// Copyright (c) 2015 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-
-part of newton;
 
 class Tolerance {
   final double distance;
