+++ packages/newton/lib/src/utils.dart
@@ -1,10 +1,8 @@
-// Copyright (c) 2015 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of newton;
-
-bool _nearEqual(double a, double b, double epsilon) =>
+bool nearEqual(double a, double b, double epsilon) =>
     (a > (b - epsilon)) && (a < (b + epsilon));
 
-bool _nearZero(double a, double epsilon) => _nearEqual(a, 0.0, epsilon);
+bool nearZero(double a, double epsilon) => nearEqual(a, 0.0, epsilon);
